requirement.txt
coincurve
pysha3